requirements.txt
# ─────────────────────────────────────────────────────────────
#  RAG Customer Support Assistant — Dependencies
#  Install: pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────

# ── Core LangChain / LangGraph ────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0

# ── Vector Database ───────────────────────────────────────────
chromadb>=0.5.0

# ── Embeddings (local — no API key needed) ───────────────────
sentence-transformers>=2.7.0

# ── PDF Processing ────────────────────────────────────────────
pypdf>=4.0.0
pdfplumber>=0.11.0

# ── API Framework ─────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9    # for file uploads
pydantic>=2.7.0

# ── Configuration ─────────────────────────────────────────────
python-dotenv>=1.0.0

# ── LLM (optional — needed only for GPT-4) ───────────────────
openai>=1.30.0

# ── PDF generation (for sample KB only) ─────────────────────
reportlab>=4.0.0

# ── Testing ───────────────────────────────────────────────────
pytest>=8.0.0
